+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [2.0.0] - 2025-05-01
+
+### Changed
+
+- Renamed Parse to ParsePath
+
+### Added
+
+- Added ParseBytes to parse changelog from bytes (issue #4)
+
 ## [1.0.0] - 2025-04-20
 
 ### Added